+++ data-raw/ordinal-sustain.R
@@ -37,6 +37,7 @@ N_CV_folds = 10L
 rerun = TRUE
 # rerun = FALSE
 plot_python = TRUE
+fig_size = c(20, 10)
 
 args = commandArgs(trailingOnly = TRUE)
 message("args = ", args |> paste(collapse = "; "))
@@ -92,9 +93,7 @@ biomarker_events_table =
     biomarker_levels,
     biomarker_groups)
 
-## ----------------------------------------------------------------------------------------------------
-#| tbl-cap: "Biomarkers used in analysis"
-#| label: "tbl-biomarker-list"
+cli::cli_inform("Biomarkers used in analysis:")
 table_out =
   biomarker_events_table |>
   select(category = biomarker_group, biomarker, levels) |>
@@ -149,6 +148,7 @@ sustain_output = run_and_save_OSA(
   dataset_name = dataset_name,
   use_parallel_startpoints = use_parallel_startpoints,
   plot = plot_python,
+  fig_size = fig_size,
   rerun = rerun,
   patient_data = patient_data,
   N_CV_folds = N_CV_folds,
@@ -170,7 +170,8 @@ sustain_output_males = run_and_save_OSA(
   dataset_name = "males",
   rerun = rerun,
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 
 
@@ -190,7 +191,8 @@ sustain_output_females = run_and_save_OSA(
   dataset_name = "females",
   rerun = rerun,
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 
 
@@ -210,7 +212,8 @@ sustain_output_cgg100plus = run_and_save_OSA(
   dataset_name = "over100",
   rerun = rerun,
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 
 
@@ -230,7 +233,8 @@ sustain_output_cgg100minus = run_and_save_OSA(
   dataset_name = "under100",
   rerun = rerun,
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 ## ----------------------------------------------------------------------------------------------------
 #| message: false
@@ -251,7 +255,8 @@ sustain_output_cgg100plus_males = run_and_save_OSA(
   dataset_name = "over100_Male",
   rerun = rerun,
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 
 
@@ -275,7 +280,8 @@ sustain_output_cgg100minus_males = run_and_save_OSA(
   dataset_name = "under100_Male",
   rerun = rerun,
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 ## ----------------------------------------------------------------------------------------------------
 #| message: false
@@ -297,7 +303,8 @@ sustain_output_cgg100plus_females = run_and_save_OSA(
   rerun = rerun,
   dataset_name = "over100_Female",
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 
 
@@ -319,6 +326,7 @@ sustain_output_cgg100minus_females = run_and_save_OSA(
   rerun = rerun,
   dataset_name = "under100_Female",
   use_parallel_startpoints = use_parallel_startpoints,
-  plot = plot_python)
+  plot = plot_python,
+  fig_size = fig_size)
 
 cli::cli_alert_info('\nEnding at: {Sys.time()}')